users-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
loguru==0.7.2